requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
Click==7.0
colorama==0.4.1
coverage==4.5.3
Flask==1.0.3
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-PyMongo==2.3.0
importlib-metadata==0.17
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
PyJWT==1.4.2
pylint==2.3.1
pymongo==3.8.0
pyparsing==2.4.0
pytest==4.6.2
pytest-dotenv==0.4.0
python-dotenv==0.10.3
six==1.12.0
typed-ast==1.3.5
wcwidth==0.1.7
Werkzeug==0.15.4
wincertstore==0.2
wrapt==1.11.1
WTForms==2.2.1
zipp==0.5.1
